STYLE: move func triggerCenterClickThrough() into src/utils
